Local/electron_moving_marker/requirement.txt
pysrt
moviepy
psycopg2-binary
PyQt5
falcon
waitress
geojson
falcon-cors